requirements.txt
Flask
Flask-PyMongo
flask-cors
Flask-JWT-Extended 
flask-bcrypt
Flask-Login
flask-restful
Flask-Limiter
nbconvert
nbformat
python-slugify
PyJWT
werkzeug
flask-debugtoolbar